CVSL-1505 add parameters needed to create licences in the backend (#399)
* CVSL-1505 add parameters needed to create licences in the backend

* CVSL-1505 add parameters to integration tests

* CVSL-1505 remove unneccessary class

+++ src/main/kotlin/uk/gov/justice/digital/hmpps/createandvaryalicenceapi/service/probation/OffenderDetail.kt
@@ -1,7 +1,7 @@
 package uk.gov.justice.digital.hmpps.createandvaryalicenceapi.service.probation
 
 data class OffenderDetail(
-  val offenderId: Int,
+  val offenderId: Long,
   val otherIds: OtherIds,
   val offenderManagers: List<OffenderManager>,
 )
